CORRECT ALL CONFIGS

+++ certbot/generate-certificate.sh
@@ -1,11 +1,17 @@
 #!/bin/bash
 
-rm -rf /etc/letsencrypt/live/coffeeinveins.ru
+mkdir -p /var/www/certbot/.well-known/acme-challenge/
 
-certbot certonly --webroot -w /var/www/certbot -d coffeeinveins.ru -d www.coffeeinveins.ru --email komy.kabachok@yandex.ru --agree-tos --non-interactive
+certbot certonly --webroot -w /var/www/certbot -d coffeeinveins.ru -d www.coffeeinveins.ru --email book.anna.book@gmail.com --agree-tos --non-interactive
 
-rm -rf /etc/nginx/cert.pem
-rm -rf /etc/nginx/key.pem
+if [ -f /etc/letsencrypt/live/coffeeinveins.ru/fullchain.pem ]; then
+    cp /etc/letsencrypt/live/coffeeinveins.ru/fullchain.pem /etc/nginx/cert.pem
+else
+    echo "Certificate not found! :(("
+fi
 
-cp /etc/letsencrypt/live/coffeeinveins.ru/fullchain.pem /etc/nginx/cert.pem
-cp /etc/letsencrypt/live/coffeeinveins.ru/privkey.pem /etc/nginx/key.pem
+if [ -f /etc/letsencrypt/live/coffeeinveins.ru/privkey.pem ]; then
+    cp /etc/letsencrypt/live/coffeeinveins.ru/privkey.pem /etc/nginx/key.pem
+else
+    echo "Private key not found! :("
+fi
